refactor: decompose brain.js router into per-domain sub-routers + thin barrel (#1832)
